+++ glean-core/android/src/main/java/mozilla/telemetry/glean/private/PingType.kt
@@ -44,12 +44,15 @@ enum class NoReasonCodes(
  *
  * @property reasonCodes The list of acceptable reason codes for this ping.
  */
+@Suppress("LongParameterList")
 class PingType<ReasonCodesEnum> (
     name: String,
     includeClientId: Boolean,
     sendIfEmpty: Boolean,
     preciseTimestamps: Boolean,
     includeInfoSections: Boolean,
+    enabled: Boolean,
+    val schedulesPings: List<String>,
     val reasonCodes: List<String>,
 ) where ReasonCodesEnum : Enum<ReasonCodesEnum>, ReasonCodesEnum : ReasonCode {
     private var testCallback: ((ReasonCodesEnum?) -> Unit)? = null
@@ -62,7 +65,9 @@ class PingType<ReasonCodesEnum> (
             sendIfEmpty = sendIfEmpty,
             preciseTimestamps = preciseTimestamps,
             includeInfoSections = includeInfoSections,
+            schedulesPings = schedulesPings,
             reasonCodes = reasonCodes,
+            enabled = enabled,
         )
     }