+++ glean-core/android/src/test/java/mozilla/telemetry/glean/GleanTest.kt
@@ -480,6 +480,8 @@ class GleanTest {
             sendIfEmpty = false,
             preciseTimestamps = true,
             includeInfoSections = true,
+            enabled = true,
+            schedulesPings = emptyList(),
             reasonCodes = listOf(),
         )
         val stringMetric = StringMetricType(
@@ -847,6 +849,8 @@ class GleanTest {
             sendIfEmpty = false,
             preciseTimestamps = true,
             includeInfoSections = true,
+            enabled = true,
+            schedulesPings = emptyList(),
             reasonCodes = listOf(),
         )
         val stringMetric = StringMetricType(
@@ -987,28 +991,15 @@ class GleanTest {
         // Set a metric configuration which will enable the telemetry.string_metric
         val metricConfig = """
             {
-              "telemetry.string_metric": true
+              "metrics_enabled": {
+                "telemetry.string_metric": true
+              }
             }
         """.trimIndent()
-        Glean.setMetricsEnabledConfig(metricConfig)
+        Glean.applyServerKnobsConfig(metricConfig)
 
         // This should result in the metric being set to "foo"
         stringMetric.set("foo")
-        assertNotNull(stringMetric.testGetValue())
-        assertEquals("foo", stringMetric.testGetValue())
-
-        // Set a metric configuration which will disable the telemetry.string_metric
-        // again, this time using the deprecated API to ensure backwards compatibility
-        val metricConfigBackwardsCompat = """
-            {
-              "telemetry.string_metric": true
-            }
-        """.trimIndent()
-        Glean.setMetricsDisabledConfig(metricConfigBackwardsCompat)
-
-        // This should not result in the metric being set to "bar", it should still
-        // contain the original "foo" string
-        stringMetric.set("bar")
         assertNotNull(stringMetric.testGetValue())
         assertEquals("foo", stringMetric.testGetValue())
     }