+++ glean-core/android/src/test/java/mozilla/telemetry/glean/debug/GleanDebugActivityTest.kt
@@ -250,6 +250,8 @@ class GleanDebugActivityTest {
             sendIfEmpty = true,
             preciseTimestamps = true,
             includeInfoSections = true,
+            enabled = true,
+            schedulesPings = emptyList(),
             reasonCodes = listOf(),
         )