+++ glean-core/ios/Glean/Metrics/Ping.swift
@@ -41,6 +41,8 @@ public class Ping<ReasonCodesEnum: ReasonCodes> {
         sendIfEmpty: Bool,
         preciseTimestamps: Bool,
         includeInfoSections: Bool,
+        enabled: Bool,
+        schedulesPings: [String],
         reasonCodes: [String]
     ) {
         self.name = name
@@ -51,6 +53,8 @@ public class Ping<ReasonCodesEnum: ReasonCodes> {
             sendIfEmpty,
             preciseTimestamps,
             includeInfoSections,
+            enabled,
+            schedulesPings,
             reasonCodes
         )
     }