+++ glean-core/ios/GleanTests/GleanTests.swift
@@ -259,6 +259,8 @@ class GleanTests: XCTestCase {
             sendIfEmpty: false,
             preciseTimestamps: true,
             includeInfoSections: true,
+            enabled: true,
+            schedulesPings: [],
             reasonCodes: []
         )
 
@@ -305,34 +307,14 @@ class GleanTests: XCTestCase {
         let metricConfigStringifiedJson =
 """
 {
-  "telemetry.counter_metric": true
+  "metrics_enabled": {
+    "telemetry.counter_metric": true
+  }
 }
 """
-        Glean.shared.setMetricsEnabledConfig(metricConfigStringifiedJson)
+        Glean.shared.applyServerKnobsConfig(metricConfigStringifiedJson)
 
         // Attempt to add to the counter, this should succeed.
-        counter.add(1)
-        if let value = counter.testGetValue() {
-            XCTAssertEqual(1, value)
-        } else {
-            XCTAssert(false, "Failed to set metric config to enable counter")
-        }
-
-        // Set a metric configuration that disables the telemetry.counter_metric
-        // again, this time using the old API to ensure backwards compatibility.
-        // The old API was inverted and mapped directly to the metrics.yaml property
-        // for each metric named 'disabled'. So in this case `true` means the metric
-        // is disabled.
-        let metricConfigBackwardsCompatible =
-"""
-{
-  "telemetry.counter_metric": false
-}
-"""
-        Glean.shared.setMetricsDisabledConfig(metricConfigBackwardsCompatible)
-
-        // Attempt to add to the counter, this should not record so the value should
-        // remain at 1
         counter.add(1)
         if let value = counter.testGetValue() {
             XCTAssertEqual(1, value)
@@ -417,6 +399,8 @@ class GleanTests: XCTestCase {
             sendIfEmpty: false,
             preciseTimestamps: true,
             includeInfoSections: true,
+            enabled: true,
+            schedulesPings: [],
             reasonCodes: []
         )