+++ glean-core/python/glean/metrics/__init__.py
@@ -25,6 +25,7 @@ from .._uniffi import StringListMetric as StringListMetricType
 # Export wrapper implementations for metric types
 from .datetime import DatetimeMetricType
 from .event import EventMetricType, EventExtras, RecordedEvent
+from .object import ObjectMetricType, ObjectSerialize
 from .labeled import (
     LabeledBooleanMetricType,
     LabeledCounterMetricType,
@@ -43,23 +44,25 @@ __all__ = [
     "CommonMetricData",
     "CounterMetricType",
     "DatetimeMetricType",
+    "EventExtras",
     "EventMetricType",
-    "QuantityMetricType",
     "LabeledBooleanMetricType",
     "LabeledCounterMetricType",
     "LabeledStringMetricType",
     "Lifetime",
     "MemoryDistributionMetricType",
     "MemoryUnit",
+    "ObjectMetricType",
+    "ObjectSerialize",
     "PingType",
+    "QuantityMetricType",
     "RecordedEvent",
-    "EventExtras",
     "RecordedExperiment",
-    "StringMetricType",
     "StringListMetricType",
-    "TimespanMetricType",
+    "StringMetricType",
     "TimeUnit",
     "TimerId",
+    "TimespanMetricType",
     "TimingDistributionMetricType",
     "UrlMetricType",
     "UuidMetricType",