+++ glean-core/python/glean/metrics/ping.py
@@ -17,6 +17,8 @@ class PingType:
         send_if_empty: bool,
         precise_timestamps: bool,
         include_info_sections: bool,
+        # enabled: bool,  -- Not currently enabled for python bindings
+        schedules_pings: List[str],
         reason_codes: List[str],
     ):
         """
@@ -32,6 +34,8 @@ class PingType:
             send_if_empty,
             precise_timestamps,
             include_info_sections,
+            True,
+            schedules_pings,
             reason_codes,
         )
         self._test_callback = None  # type: Optional[Callable[[Optional[str]], None]]