+++ gradle-plugin/src/main/groovy/mozilla/telemetry/glean-gradle-plugin/GleanGradlePlugin.groovy
@@ -50,10 +50,10 @@ abstract class GleanMetricsYamlTransform implements TransformAction<TransformPar
 @SuppressWarnings("GrPackage")
 class GleanPlugin implements Plugin<Project> {
     // The version of glean_parser to install from PyPI.
-    private String GLEAN_PARSER_VERSION = "13.0"
+    private String GLEAN_PARSER_VERSION = "14.0"
     // The version of Miniconda is explicitly specified.
     // Miniconda3-4.5.12 is known to not work on Windows.
-    private String MINICONDA_VERSION = "4.5.11"
+    private String MINICONDA_VERSION = "24.3.0-0"
 
     private String TASK_NAME_PREFIX = "gleanGenerateMetrics"
 
@@ -480,9 +480,9 @@ except:
 
                 // If we have a git package (a la `git+https://github.com`) we install that.
                 if (parserVersion.matches("git.+")) {
-                    conda "Miniconda3", "Miniconda3-${MINICONDA_VERSION}", "64", [parserVersion]
+                    conda "Miniconda3", "Miniconda3-py311_${MINICONDA_VERSION}", "64", [parserVersion]
                 } else {
-                    conda "Miniconda3", "Miniconda3-${MINICONDA_VERSION}", "64", ["glean_parser~=${parserVersion}"]
+                    conda "Miniconda3", "Miniconda3-py311_${MINICONDA_VERSION}", "64", ["glean_parser~=${parserVersion}"]
                 }
             }
             File envDir = new File(
@@ -552,7 +552,7 @@ except:
     void apply(Project project) {
         isOffline = project.gradle.startParameter.offline
 
-        project.ext.glean_version = "59.0.0"
+        project.ext.glean_version = "60.0.0"
         def parserVersion = gleanParserVersion(project)
 
         // Print the required glean_parser version to the console. This is