+++ tools/embedded-uniffi-bindgen/src/lib.rs
@@ -1,0 +1,1 @@
+// intentionally left empty. only exists to make a warning go away